export toLoggingerror
