+++ packages/logging-wrapper/src/index.ts
@@ -4,6 +4,10 @@ import {
   initializeLoggingHooks as fastifyLoggingHooks,
 } from "./fastify-logging-wrapper.js";
 import { DestinationStream } from "pino";
+import {
+  toLoggingError as baseToLoggingError,
+  LoggingError as baseLoggingError,
+} from "./logging-wrapper-entities.js";
 
 export const getLoggingConfiguration = (
   loggerDestination?: DestinationStream,
@@ -11,3 +15,6 @@ export const getLoggingConfiguration = (
 
 export const initializeLoggingHooks = (server: FastifyInstance): void =>
   fastifyLoggingHooks(server);
+
+export const toLoggingError = baseToLoggingError;
+export type LoggingError = baseLoggingError;
