+++ packages/logging-wrapper/src/logging-wrapper-entities.ts
@@ -1,3 +1,6 @@
+import { FastifyError } from "fastify";
+import { LifeEventsError, isLifeEventsError } from "shared-errors";
+
 export interface LoggingRequest {
   scheme: string;
   method: string;
@@ -61,3 +64,55 @@ export const REDACTED_PATHS = [
 export const MESSAGE_KEY = "message";
 
 export const REQUEST_ID_LOG_LABEL = "request_id";
+
+const UNHANDLED_EXCEPTION_CODE = "UNHANDLED_EXCEPTION";
+
+export const toLoggingError = (
+  error: LifeEventsError | FastifyError,
+): LoggingError => {
+  const output = {
+    class: parseErrorClass(error),
+    message: error.message,
+    trace: error.stack,
+  };
+
+  if (isLifeEventsError(error)) {
+    return { ...output, code: error.name, process: error.errorProcess };
+  }
+
+  return {
+    ...output,
+    code: error.code ?? UNHANDLED_EXCEPTION_CODE,
+  };
+};
+
+const parseErrorClass = (
+  error: FastifyError | LifeEventsError,
+): LogErrorClasses => {
+  let statusCode = isLifeEventsError(error)
+    ? error.errorCode
+    : error.statusCode;
+  if (!statusCode) {
+    return LogErrorClasses.UnknownError;
+  }
+
+  statusCode = Number(statusCode);
+
+  if (statusCode === 502) {
+    return LogErrorClasses.GatewayError;
+  }
+
+  if (statusCode >= 500) {
+    return LogErrorClasses.ServerError;
+  }
+
+  if (statusCode === 422) {
+    return LogErrorClasses.ValidationError;
+  }
+
+  if (statusCode >= 400) {
+    return LogErrorClasses.RequestError;
+  }
+
+  return LogErrorClasses.UnknownError;
+};
